requirements.txt
pygame==2.5.2
python-socketio>=5.0.0
aiohttp>=3.8.0
python-engineio==4.8.0
eventlet>=0.33.0
requests>=2.32.0
websocket-client>=1.8.0 